refactor: add mvc register users
